+++ app.js
@@ -1,7 +1,7 @@
 import express from "express";
 import cookieParser from "cookie-parser";
 import session from "express-session";
-import routes from "./routes/index.js";
+import routes from "./app/index.js";
 import passport from "passport";
 import cors from "cors";
 import "./strategy/local-strategy.js";
